feat: 🎸 fetch active listings instead of featured listings
BREAKING CHANGE: All active listings are now fetched instead of just featured ones. To
reflect this, the `featuredEtsyListing` node has been renamed to
`etsyListing` (and `allFeaturedEtsyListing` is now `allEtsyListing`).

Featured listings can be queried in the GraphQL playground by querying for listings
with a featured_rank value that is not null.